WIP - moved SamCliInfoResponse
