+++ src/shared/sam/cli/samCliInvokerUtils.ts
@@ -11,22 +11,9 @@ import { stat } from '../../filesystem'
 import { ChildProcessResult } from '../../utilities/childProcess'
 import { SamCliVersionValidatorResult } from './samCliVersionValidator'
 
-/**
- * Maps out the response text from the sam cli command `sam --info`
- */
-// TODO : Move this back to samCliInfo.ts
-export interface SamCliInfoResponse {
-    version: string
-}
-
 export interface SamCliProcessInvoker {
     invoke(options: SpawnOptions, ...args: string[]): Promise<ChildProcessResult>
     invoke(...args: string[]): Promise<ChildProcessResult>
-}
-
-export interface SamCliProcessInfo {
-    info?: SamCliInfoResponse
-    lastModified?: Date
 }
 
 export interface SamCliUtils {
